Improve steering.

+++ esp32/main/motor.h
@@ -29,7 +29,7 @@ private:
 // more of the joystick's range to pivot actions.
 const int PIVOT = 35;
 
-void compute_power(int rx, int ry, int& power_left, int& power_right, int pivot = PIVOT, int max_power = 100);
+void compute_power(int rx, int ry, int& power_left, int& power_right, double pivot = PIVOT, int max_power = 100);
 
 // -1 to +1
 void set_motors(double m1, double m2);
